Add first stack navigator
